More toggling abilities
Inactive toggles work for interior territories.
Also added the option to flip the toggle only if the inactive toggle
was clicked (basically from 2 to 1).
Will add the ability to switch from 1 to 2 later in the evening.
Also will add the ability to add a toggle between an interior and
exterior.