More toggling toggles
You can now toggle anything as long as it involves two of the same
faction-owned territories.
I couldn’t find any toggle-related bugs so this should be good for the
alpha.
My check for toggle loops results in an infinite loop…. This is not
really urgent so I’ll look into it later. Even with toggle infinite
loops, the program doesn’t crash so it’s fine for now.