Toggle anything but loops & no remaining borders
You should be able toggle anything you want. However, if the AI detects
that a loop might occur because of it, it will say “loop for <territory
# that is part of the loop>” and not do the toggle. I still don’t know
how to solve the issue that sometimes you have to click on a toggle
multiple times to register. (maybe missing the hitbox?)

Another instance of it not toggling is for example, all 6 of the
borders are used in 1 territory (so 5 toggles going in and 1 going out
of the territory). You can toggle any of the 5 going in but you can’t
toggle the one going out because the AI isn’t smart enough to toggle
more than 2 territories per click. I’ll look into making the AI more
sophisticated in the afternoon.

I am still tinkering around with battle toggles that result in one
faction either cornering another faction’s territories or circling them
completely. I’m close but I’ll release a polished version later today.

Let me know if you encounter any bugs/concerns for the toggles apart
from the ones I already said will be solved soon.